+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Functionality to enable/disable specific menu node (#85, DEV-65561)
 
+### Removed
+- Menu admin edit form save and continue edit button (#69112)
+
 ## [2.12.0] - 2020-08-24
 ### Added
 - PHP 7.4 support (#118)